+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Changed
 
-
+- Vectorized *Mix* nodes.
 
 ## 2.2.0 (01 September 2020)